# perf(storage_proofs): Allocate outside the parents() call