In order to save some allocations, the vector for parents is
allocated outside the call. This is on optimisation for loops
where that vector can be re-used in several parents() calls.

It's not clear that it really improves the performance. When
running the benchmark for encoding a single node, there's no
significant difference when this change is applied:

```
group                             HEAD2                                  less-alloc2
-----                             -----                                  -----------
encode single node/Blake2s/10     1.00     13.1±0.63µs        ? B/sec    1.01     13.2±0.62µs        ? B/sec
encode single node/Blake2s/3      1.00     11.8±0.52µs        ? B/sec    1.07     12.7±0.59µs        ? B/sec
encode single node/Blake2s/5      1.00     12.3±0.92µs        ? B/sec    1.05     12.8±0.77µs        ? B/sec
encode single node/Pedersen/10    1.00     11.8±0.61µs        ? B/sec    1.00     11.8±0.59µs        ? B/sec
encode single node/Pedersen/3     1.02     11.5±0.69µs        ? B/sec    1.00     11.3±0.47µs        ? B/sec
encode single node/Pedersen/5     1.00     11.5±0.65µs        ? B/sec    1.00     11.5±0.93µs        ? B/sec
encode single node/Sha256/10      1.00     12.0±0.77µs        ? B/sec    1.02     12.2±1.22µs        ? B/sec
encode single node/Sha256/3       1.01     11.3±0.52µs        ? B/sec    1.00     11.2±0.55µs        ? B/sec
encode single node/Sha256/5       1.00     11.4±0.56µs        ? B/sec    1.00     11.4±0.63µs        ? B/sec
```

Closes #575.