@porcuquine 2019-04-13
Thanks, @vmx.

Given that this doesn't show a performance improvement  and makes the code harder to reason about, I'm inclined not to make the change. I don't think we should avoid straightforward value-oriented/functional style except when it carries a significant performance benefit. #575 hypothesized that this would be the case here, but you've falsified that hypothesis.

My preference would be to close #575 with a reference to your benchmarks and the relevant code here. So your effort wasn't wasted, even if the knowledge it produced doesn't make it into the code. (We can probably learn from this example and not generally worry too much about ephemeral allocation of small vectors.)

What do you think?