@vmx 2019-04-13
I fully agree. I'm just wondering if

 - we miss anything obvious why this isn't faster
 - the benchmark is no good. One problem I see with the benchmark is that it really has a tiny runtime. I'm unsure if you can actually measure meaning full things at that scale or if basically everything is just noise.