@porcuquine 2019-04-13
You make a good point. Specifically, it's not clear to me (after a brief look) that the benchmark is even calling the function you changed here. As far as I can tell, your speculative optimization was to `parents`, but the benchmark seems to be pre-generating that data without even calling the function.

I may be restating your second point.

So maybe the thing to do is modify the benchmark to actually exercise these changes and see whether that improves the results.