@dignifiedquire 2019-04-15
Agreed with what has been said, the optimized path is not actually called inside the referenced benchmark.