@dignifiedquire 2019-04-15
Maybe simply adding a benchmark which just measures the performance of `parents` would be good