@vmx 2019-04-15
I totally missed that I'm not actually benchmarking this thing. So I agree that adding another benchmark makes sense, perhaps one that does something similar to `vde::encode()` (or even just calls `vde::encode()`). I'm happy to work on the benchmark, though it will wait until Friday in two weeks.