@dignifiedquire 2019-04-16
I have a local benchmark trying some things out, and this gives a speed increase of ~30% when using max caching for a single node encoding. So this is def worth it