@vmx 2019-05-03
I've added a new benchmark for measuring `parents()` only. I don't see any difference in performance, I wonder if I did something wrong. The new numbers are ("after" is with this change, "before" is without this change):

```
    group                              after                                  before
    -----                              -----                                  ------
    parents in a loop/Blake2s/10       1.00    135.4±4.02µs        ? B/sec    1.01    136.5±8.78µs        ? B/sec
    parents in a loop/Blake2s/1000     1.00      9.1±0.27ms        ? B/sec    1.01      9.2±0.18ms        ? B/sec
    parents in a loop/Blake2s/50       1.00   563.5±31.95µs        ? B/sec    1.00   561.1±17.98µs        ? B/sec
    parents in a loop/Pedersen/10      1.00    134.6±4.32µs        ? B/sec    1.02    137.1±4.35µs        ? B/sec
    parents in a loop/Pedersen/1000    1.00      9.0±0.18ms        ? B/sec    1.01      9.2±0.19ms        ? B/sec
    parents in a loop/Pedersen/50      1.01   560.3±26.94µs        ? B/sec    1.00   554.1±19.80µs        ? B/sec
    parents in a loop/Sha256/10        1.00    133.6±3.03µs        ? B/sec    1.00    133.9±4.52µs        ? B/sec
    parents in a loop/Sha256/1000      1.00      9.1±0.22ms        ? B/sec    1.02      9.2±0.23ms        ? B/sec
    parents in a loop/Sha256/50        1.00   549.7±15.63µs        ? B/sec    1.01   553.0±15.38µs        ? B/sec
```