@vmx 2019-05-03
@dignifiedquire pointed out that I should run with `FIL_PROOFS_MAXIMIZE_CACHING=true`.

Now there is a significant difference in performance:

```
group                              after                                   before
-----                              ------                                  -------
parents in a loop/Blake2s/10       1.00      5.1±0.51µs        ? B/sec     1.33      6.8±0.58µs        ? B/sec
parents in a loop/Blake2s/1000     1.00 1011.4±694.48µs        ? B/sec     1.10 1109.6±509.07µs        ? B/sec
parents in a loop/Blake2s/50       1.00     30.8±3.06µs        ? B/sec     1.32     40.7±5.97µs        ? B/sec
parents in a loop/Pedersen/10      1.00      5.2±0.48µs        ? B/sec     1.37      7.1±0.72µs        ? B/sec
parents in a loop/Pedersen/1000    1.00  939.9±619.75µs        ? B/sec     1.17 1101.9±517.17µs        ? B/sec
parents in a loop/Pedersen/50      1.00     31.2±3.57µs        ? B/sec     1.36     42.5±3.81µs        ? B/sec
parents in a loop/Sha256/10        1.00      4.9±0.31µs        ? B/sec     1.37      6.7±0.50µs        ? B/sec
parents in a loop/Sha256/1000      1.00  918.5±560.68µs        ? B/sec     1.25 1149.1±549.77µs        ? B/sec
parents in a loop/Sha256/50        1.00     31.0±3.35µs        ? B/sec     1.34     41.6±4.80µs        ? B/sec
```