@vmx 2019-05-03
Changing the for loops to iterators doesn't have a performance impact.